refactor: extract ripple helper
